ROCKLINUX: ld: allow building on x86-64
Original description below:

Allows bin86 to compile on x86_64, which permits lilo to
be used in a 64-bit system.  I'm dubious about ROCK's attempts to
automatically add dual-licensing to all their patches, but bin86 is
already GPL'd.  I've heard from the maintainer that this will be
included in the next upload.

Signed-off-by: Ken Moffat <ken@kenmoffat.uklinux.net>

Link: https://aur.archlinux.org/cgit/aur.git/tree/bin86-0.16.17-x86_64-1.patch?h=bin86&id=7f8f84df91e38c08f9fbb820592afa17215e9e48
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>